+++ routes/api/contacts.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const createError = require("http-errors");
-const joi = require('joi');
 const router = express.Router();
 
+const { addContactsValidation, putContactsValidation } = require("../../middlewares/joiValidation");
 const contacts = require("../../models/contacts");
 
 
@@ -33,29 +33,8 @@ router.get('/:id', async (req, res, next) => {
 });
 
 
-router.post('/', async (req, res, next) => {
+router.post('/', addContactsValidation, async (req, res, next) => {
   try {
-    const schema = joi.object({
-      name: joi.string()
-        .min(2)
-        .max(30)
-        .required(),
-      email: joi.string()
-        .min(5)
-        .max(50)
-        .required(),
-      phone: joi.string()
-        .min(8)
-        .max(16)
-        .required(),
-    });
-
-    const validationResult = schema.validate(req.body);
-    if (validationResult.error) {
-      console.log(validationResult.error)
-      return res.json({ message: "missing required name field", status: 400 });
-    }
-
     const { name, email, phone } = req.body;
     const data = await contacts.addContact(name, email, phone);
     res.json({ data, status: 201 });
@@ -82,13 +61,10 @@ router.delete('/:id', async (req, res, next) => {
 });
 
 
-router.put('/:id', async (req, res, next) => {
+router.put('/:id', putContactsValidation, async (req, res, next) => {
   try {
     const { id } = req.params;
     const { name, email, phone } = req.body;
-    if (!name || !email || !phone) {
-      return res.json({ message: "missing fields", status: 400 });
-    }
     const data = await contacts.updateContact(id, name, email, phone);
     if (!data) {
       const error = createError(404, "Not found");
